Add server-client communication
This sets up the server's internal state (a list of connected clients,
a representation of the secret, which letter of the secret we're on) and
communication between the server and client. The client now mostly just
passes off "move" events to the server and updates the display according
to what the server sends back. When the secret letter changes, the
server sends all clients a message if their room color has changed.

The client side now just uses `client.js`, and the HTML and CSS have
been updated to make things a little easier on the client and to adapt
(I hope) better to different-sized displays.

I switched off of CSS variables because I was worried about support on
various mobile browsers after consulting the MDN compatibility table.
But perhaps it would have been fine. The caniuse table looks pretty
promising.