TEMP: Add server-client communication
This is neither tested nor done, so expect it to be rebased. I'm mostly
just pushing to avoid losing progress. The work so far goes _most_ of
the way, as a rough draft, toward setting up the server's internal state
and communication between the client and server, but we aren't actually
instantiating anything on the client side at the moment, so we aren't
yet establishing communication.