Add prompt for video playback
Boooooooo. Mobile devices don't allow JavaScript to play videos
automatically; they can only start playing a video in the handler for an
user-initiated event like a click. This means that we can't just start
playing the video automatically, and instead we have to show all
connected users a prompt that plays the video when they click it. The
prompt tries to hint to users that they should click at the same time.